svg/painting/parsing/fill-rule-valid.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="800"><title>SVG Painting: parsing fill-rule with valid values</title><metadata><h:link href="https://svgwg.org/svg2-draft/painting.html#FillRuleProperty" rel="help"/><h:meta content="fill-rule supports the full grammar 'nonzero | evenodd'." name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/parsing-testcommon.js"/><script><![CDATA[

test_valid_value("fill-rule", "nonzero");
test_valid_value("fill-rule", "evenodd");

  ]]></script></svg>
